feat(me): split into components
