+++ src/components/Contact.tsx
@@ -1,0 +1,21 @@
+import React from 'react'
+
+export default function Contact(): JSX.Element {
+  return (
+    <div>
+      <table>
+        <tr>
+          <td>
+            <h1>Discord</h1>
+          </td>
+          <td>
+            <h1>Email</h1>
+          </td>
+          <td>
+            <h1>Linkedin</h1>
+          </td>
+        </tr>
+      </table>
+    </div>
+  )
+}
