+++ src/components/MeAsAPerson.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+export default function MeAsAPerson(): JSX.Element {
+  return (
+    <div>
+      <h1>ME AS A PERSON</h1>
+      <p>Hello, my name is Alexandre Franquet, self-nicknamed Miouzora, I’m 20. I’m a french student, who’s currently studying at Epitech school. I’m a developper who works to be a game developer one day.</p>
+    </div>
+  )
+}
