+++ src/components/PersoExp.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+import PersoList from './PersoList'
+
+export default function PersoExp(): JSX.Element {
+  return (
+    <div>
+      <h1>PERSONAL EXPERIENCE</h1>
+      <PersoList />
+    </div>
+  )
+}
