+++ src/components/ProExp.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+import ProList from './ProList'
+
+export default function ProExp(): JSX.Element {
+  return (
+    <div>
+      <h1>PROFESSIONAL EXPERIENCE</h1>
+      <ProList />
+    </div>
+  )
+}
